New Updated Database file!
Fixed redundant columns in the tables such as is_Verified from user_token table which can now only be found in the user table to check if the user has verified their email. Also added a new table shipping_address and removed the shipping_address column in the orders table to normalize the tables.